+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/HasAttributes.kt
@@ -56,6 +56,10 @@ public interface HasAttributes {
 
     /**
      * Set a collection of values.
+     * Set a collection of primitive values as an attribute.
+     * Every element should be of the same type, and only String, Integer, Long, Double and Boolean
+     * are supported.
+     * If the collection contains mixed values, null, or any other type: the behavior is undefined.
      *
      * @param name the attribute name
      * @param value the value to set the attribute to
@@ -77,8 +81,8 @@ public interface HasAttributes {
      * @param value the value to set the attribute to
      */
     public fun setAttribute(name: String, value: LongArray?)
-
-    /**
+    
+     /**
      * Set an array of double values.
      *
      * @param name the attribute name