+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/SpanImpl.kt
@@ -184,9 +184,7 @@ public class SpanImpl internal constructor(
 
     override fun setAttribute(name: String, value: Collection<Any>?) {
         if (!isEnded()) {
-            if (value != null) {
-                attributes[name] = value
-            }
+            attributes[name] = value
         }
     }