+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## TBC
+
+### Enhancements
+
+* Custom attributes can now be set on a span, including as arrays of primitives (int, long, double, boolean, string).
+  [#252](https://github.com/bugsnag/bugsnag-android-performance/pull/252)
+
 ## 1.4.0 (2024-06-27)
 
 ### Enhancements